Floor syncing & Usb Controller added
Added floor sync functionality. (works for setting debug grid size too)
Added WIP Usb controller. (certain reference included, need to test if it works on another workstation)
Reworked some Int32s into Int16s to better match hardware and save on memory. 
Added safety to retrieving some values. 
Added some utility extension methods.
Removed some unused code and resources (needs a second global pass).

+++ LightSpace_WPF_Engine/LightSpace_WPF_Engine/Models/Utility/DataTool.cs
@@ -100,7 +100,7 @@ namespace LightSpace_WPF_Engine.Models.Utility
         /// <param name="tileNumberBits"> Tile number as needed by hardware. </param>
         /// <param name="colors"> 64 separate bytes giving a color of every led on the tile. </param>
         /// <returns> Returns byte array containing the given parameters ready to send colors to specified hardware tile. </returns>
-        public static byte[] InputDataToByteArray(int tileNumberBits, Color[,] colors)
+        public static byte[] InputDataToByteArray(short tileNumberBits, Color[,] colors)
         {
             //TODO: 00 Test if this method is functional
             // create a byte array sized 66 bytes long
